# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flamoral Dating Platform API
  description: |
    Complete API documentation for the Flamoral Dating Platform.

    This dating platform features:
    - User authentication and profiles
    - Discovery and matching system
    - Real-time messaging
    - Subscription and coins system
    - Boost products
    - Privacy controls
    - Notifications
    - Payment processing
  version: 1.0.0
  contact:
    name: Flamoral Platform Team
    email: platform@flamoral.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.flamoral.com
    description: Production API
...
paths:
...
  /api/v1/conversations:
    get:
      tags: [Conversations]
      summary: Get conversations
      operationId: getConversations
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedConversations'
  /api/v1/conversations/{id}:
    get:
      tags: [Conversations]
      summary: Get conversation details
      operationId: getConversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /api/v1/conversations/{id}/messages:
    get:
      tags: [Conversations]
      summary: Get conversation messages
      operationId: getConversationMessages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Conversation messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMessages'
  /api/v1/conversations/{id}/read:
    put:
      tags: [Conversations]
      summary: Mark conversation as read
      operationId: markConversationRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as read
...
  /api/v1/messages:
    post:
      tags: [Messages]
      summary: Send message
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /api/v1/messages/unread-count:
    get:
      tags: [Messages]
      summary: Get unread message count
      operationId: getUnreadMessageCount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
  /api/v1/messages/{messageId}:
    get:
      tags: [Messages]
      summary: Get message
      operationId: getMessage
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      tags: [Messages]
      summary: Update message
      operationId: updateMessage
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Message updated
    delete:
      tags: [Messages]
      summary: Delete message
      operationId: deleteMessage
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message deleted
  /api/v1/messages/{messageId}/status:
    put:
      tags: [Messages]
      summary: Update message status
      operationId: updateMessageStatus
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [delivered, read]
      responses:
        '200':
          description: Status updated

  # ============== NOTIFICATION SERVICE ==============
...
  /api/v1/media/video:
    post:
      tags: [Media]
      summary: Upload video
      operationId: uploadVideo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
      responses:
        '201':
          description: Video uploaded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Photo:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        isPrimary:
          type: boolean
        order:
          type: integer
...
    DiscoveryProfile:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        age:
          type: integer
        bio:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        distance:
          type: number
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        participant:
          $ref: '#/components/schemas/DiscoveryProfile'
        lastMessage:
          $ref: '#/components/schemas/Message'
        unreadCount:
          type: integer
    PaginatedConversations:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, image, gif, voice]
        status:
          type: string
          enum: [sent, delivered, read]
        createdAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [conversationId, content]
      properties:
        conversationId:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, image, gif, voice]
    PaginatedMessages:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer